iothub: splitting the files to match the convention
